refactor(media): consolidate media components under src/lib/media
Rename src/lib/images to src/lib/media and move image-carousel,
media-scroller, fullscreen-media-viewer, and media-instructions-editor
into per-component sub-directories alongside image-handler,
image-proposal-modal, and ai-image-enhancer. Update all consumer
imports and internal relative paths accordingly.
