+++ src/lib/media/ai-image-enhancer/ai-image-enhancer.js
