+++ src/lib/media/fullscreen-media-viewer/fullscreen-media-viewer.js
