+++ src/lib/media/image-carousel/image-carousel.js
