+++ src/lib/media/media-scroller/media-scroller.js
